pyproject.toml
[project]
name = "ninja-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai>=2.15.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
]

[project.scripts]
ninjacli = "ninjacli.main:main"
